add UserInGroup to SubjectsController

+++ HomewOurK.WebAPI/Controllers/SubjectsController.cs
@@ -5,6 +5,7 @@ using HomewOurK.WebAPI.Helpers;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Text.RegularExpressions;
 
 namespace HomewOurK.WebAPI.Controllers
 {
@@ -37,42 +38,72 @@ namespace HomewOurK.WebAPI.Controllers
 					return Ok(subjects);
 				return NotFound("No subjects was found for the group with id = " + groupId);
 			}
-			return BadRequest();
+			return Unauthorized();
 		}
 
 		[HttpGet("GetSubject")]
 		public IActionResult GetSubject(int subjectId, int groupId)
 		{
-			var subject = _subjectService.GetSubjectById(subjectId, groupId);
+			var email = CookieHelper.GetEmailByCookie(_httpContextAccessor);
 
-			if (subject != null)
-				return Ok(subject);
-			return NotFound($"The subject wasn't found");
+			if (_userService.UserInGroup(groupId, email))
+			{
+				var subject = _subjectService.GetSubjectById(subjectId, groupId);
+
+				if (subject != null)
+					return Ok(subject);
+				return NotFound($"The subject wasn't found");
+			}
+			return Unauthorized();
 		}
 
 		[HttpPost]
 		public IActionResult AddSubject(Subject subject)
 		{
-			subject.Id = 0;
-			if (_subjectService.AddSubject(subject))
-				return Ok(subject);
-			return BadRequest("The subject has not been added");
+			var email = CookieHelper.GetEmailByCookie(_httpContextAccessor);
+
+			if (_userService.UserInGroup(subject.GroupId, email))
+			{
+				subject.Id = 0;
+				if (_subjectService.AddSubject(subject))
+					return Ok(subject);
+				return BadRequest("The subject has not been added");
+			}
+			return Unauthorized();
 		}
 
 		[HttpPatch]
 		public IActionResult UpdateSubject(Subject subject)
 		{
-			if (_subjectService.UpdateSubject(subject))
-				return Ok(subject);
-			return BadRequest("The subject has not been updated");
+			var email = CookieHelper.GetEmailByCookie(_httpContextAccessor);
+
+			if (_userService.UserInGroup(subject.GroupId, email))
+			{
+				if (_subjectService.UpdateSubject(subject))
+					return Ok(subject);
+				return BadRequest("The subject has not been updated");
+			}
+			return Unauthorized();
 		}
 
 		[HttpDelete]
-		public IActionResult DeleteSubject(Subject subject)
+		public IActionResult DeleteSubject(int subjectId, int groupId)
 		{
-			if (_subjectService.DeleteSubject(subject))
-				return Ok(subject);
-			return BadRequest("The subject has not been deleted");
+			var email = CookieHelper.GetEmailByCookie(_httpContextAccessor);
+
+			if (_userService.UserInGroup(groupId, email))
+			{
+				var subject = new Subject
+				{
+					Name = "deleted subject",
+					GroupId = groupId,
+					Id = subjectId
+				};
+				if (_subjectService.DeleteSubject(subject))
+					return Ok(subject);
+				return BadRequest("The subject has not been deleted");
+			}
+			return Unauthorized();
 		}
 	}
 }
